requirements.txt
appnope==0.1.0
backcall==0.2.0
certifi==2020.6.20
cloudpickle==1.6.0
colorama==0.4.3
colour-science==0.3.15
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
ImageTOText==0.0.1
imutils==0.5.3
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
jellyfish==0.8.2
joblib==0.17.0
jupyter-client==6.1.7
jupyter-core==4.6.3
kiwisolver==1.3.1
matplotlib==3.3.2
ml-metrics==0.1.4
networkx==2.5
numpy==1.19.2
opencv-python==4.4.0.44
pandas==1.1.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.7
ptyprocess==0.6.0
Pygments==2.7.1
pyparsing==2.4.7
pytesseract==0.3.6
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
pyzmq==19.0.2
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
threadpoolctl==2.1.0
tifffile==2020.10.1
tornado==6.0.4
tqdm==4.51.0
traitlets==5.0.4
wcwidth==0.2.5
pywin32==228